Move from long to int part 3.

+++ src/contrib/persistence/Akka.Persistence.Query.InMemory/CurrentPersistenceIdsPublisher.cs
@@ -135,7 +135,7 @@ namespace Akka.Persistence.Query.InMemory
             return Actor.Props.Create(() => new LivePersistenceIdsPublisher(refreshInterval, writeJournalPluginId));
         }
 
-        private long _lastOrderingOffset;
+        private int _lastOrderingOffset;
         private readonly ICancelable _tickCancelable;
         private readonly IActorRef _journalRef;
         private readonly DeliveryBuffer<string> _buffer;
